Check in progress towards the next release
+ Added TString for safe string manipulation
+ big fixes and improvements to HTTP, internet, and ESP8266 system
+ Added cyg_profile_func_* support to detect for stack/heap collisions
+ Added direct support for memory monitoring removing dependency on
RamMonitor.h
+ Added module debugging and logging support
+ Added support for network time via timeapi.org
+ Lots of misc bug fixes

+++ ELDisplay.h
@@ -814,7 +814,7 @@ private:
 	char*
 	ReallocString(
 		char*	inStr,
-		int		inNewSize);
+		size_t	inNewSize);
 
 	void
 	FreeString(
